@@ -3,16 +3,16 @@ class PasswordsController < Devise::PasswordsController
     self.resource = resource_class.send_reset_password_instructions(resource_params)
     yield resource if block_given?
     if successfully_sent?(resource)
-      render(:json => {:success => true})
+      render(json: {success: true})
     else
-      render(:json => {:errors => resource.errors}, :status => 500)
+      render(json: {errors: resource.errors}, status: 500)
     end
   end
 
   def edit
     self.resource = resource_class.new
     resource.reset_password_token = params[:reset_password_token]
-    render("edit", :layout => "info_window")
+    render('edit', layout: 'info_window')
   end
 
   def update
@@ -22,7 +22,7 @@ class PasswordsController < Devise::PasswordsController
       resource.unlock_access! if unlockable?(resource)
       sign_in(resource_name, resource)
     end
-    redirect_to(:controller => "main", :action => "index")
+    redirect_to(controller: 'main', action: 'index')
   end
 
 private
