@@ -7,9 +7,9 @@ class RemindersController < ApplicationController
     if @reminder.save
       ThingMailer.reminder(@reminder.thing).deliver
       @reminder.update_attribute(:sent, true)
-      render(:json => @reminder)
+      render(json: @reminder)
     else
-      render(:json => {:errors => @reminder.errors}, :status => 500)
+      render(json: {errors: @reminder.errors}, status: 500)
     end
   end
 
