@@ -1,5 +1,5 @@
 class SessionsController < Devise::SessionsController
-  skip_before_filter :verify_authenticity_token, :only => [:destroy]
+  skip_before_action :verify_authenticity_token, only: [:destroy]
 
   def new
     redirect_to(root_path)
@@ -10,9 +10,9 @@ class SessionsController < Devise::SessionsController
     if resource
       sign_in(resource_name, resource)
       yield resource if block_given?
-      render(:json => resource)
+      render(json: resource)
     else
-      render(:json => {:errors => {:password => [t("errors.password")]}}, :status => 401)
+      render(json: {errors: {password: [t('errors.password')]}}, status: 401)
     end
   end
 
@@ -21,6 +21,6 @@ class SessionsController < Devise::SessionsController
     sign_out(resource_name) if signed_in
     Devise.sign_out_all_scopes ? sign_out : sign_out(resource_name)
     yield resource if block_given?
-    render(:json => {:success => signed_in})
+    render(json: {success: signed_in})
   end
 end
