@@ -1,19 +1,18 @@
 class UsersController < Devise::RegistrationsController
   def edit
-    render("sidebar/edit_profile", :layout => "sidebar")
+    render('sidebar/edit_profile', layout: 'sidebar')
   end
 
   def update
     self.resource = resource_class.to_adapter.get!(send(:"current_#{resource_name}").to_key)
-    prev_unconfirmed_email = resource.unconfirmed_email if resource.respond_to?(:unconfirmed_email)
     if update_resource(resource, account_update_params)
       yield resource if block_given?
-      sign_in(resource_name, resource, :bypass => true)
-      flash[:notice] = "Profile updated!"
-      redirect_to(:controller => "sidebar", :action => "search")
+      sign_in(resource_name, resource, bypass: true)
+      flash[:notice] = 'Profile updated!'
+      redirect_to(controller: 'sidebar', action: 'search')
     else
       clean_up_passwords(resource)
-      render(:json => {:errors => resource.errors}, :status => 500)
+      render(json: {errors: resource.errors}, status: 500)
     end
   end
 
@@ -22,10 +21,10 @@ class UsersController < Devise::RegistrationsController
     if resource.save
       yield resource if block_given?
       sign_in(resource_name, resource)
-      render(:json => resource)
+      render(json: resource)
     else
       clean_up_passwords(resource)
-      render(:json => {:errors => resource.errors}, :status => 500)
+      render(json: {errors: resource.errors}, status: 500)
     end
   end
 
@@ -38,5 +37,4 @@ private
   def account_update_params
     params.require(:user).permit(:address_1, :address_2, :city, :current_password, :email, :name, :organization, :password, :password_confirmation, :remember_me, :sms_number, :state, :voice_number, :zip)
   end
-
 end
