@@ -4,11 +4,6 @@ class ThingMailer < ActionMailer::Base
   def reminder(thing)
     @thing = thing
     @user = thing.user
-    mail(
-      {
-        :to => thing.user.email,
-        :subject => ["Remember to shovel", thing.name].compact.join(' '),
-      }
-    )
+    mail(to: thing.user.email, subject: ['Remember to shovel', thing.name].compact.join(' '))
   end
 end
