@@ -6,14 +6,14 @@ class ThingsControllerTest < ActionController::TestCase
   end
 
   test 'should list hydrants' do
-    get :show, :format => 'json', :lat => 42.358431, :lng => -71.059773
+    get :show, format: 'json', lat: 42.358431, lng: -71.059773
     assert_not_nil assigns :things
     assert_response :success
   end
 
   test 'should update hydrant' do
     assert_not_equal 'Birdsill', @thing.name
-    put :update, :format => 'json', :id => @thing.id, :thing => {:name => 'Birdsill'}
+    put :update, format: 'json', id: @thing.id, thing: {name: 'Birdsill'}
     @thing.reload
     assert_equal 'Birdsill', @thing.name
     assert_not_nil assigns :thing
