@@ -2,6 +2,6 @@ class Address
   include Geokit::Geocoders
 
   def self.geocode(address)
-    MultiGeocoder.geocode(address).ll.split(',').map{|s| s.to_f}
+    MultiGeocoder.geocode(address).ll.split(',').collect(&:to_f)
   end
 end
