@@ -8,7 +8,7 @@
 # You can use `rake secret` to generate a secure secret key.
 
 if Rails.env.production? && ENV['SECRET_TOKEN'].blank?
-  raise 'The SECRET_TOKEN environment variable is not set.\n
+  fail 'The SECRET_TOKEN environment variable is not set.\n
     To generate it, run "rake secret", then set it with "heroku config:set SECRET_TOKEN=the_token_you_generated"'
 end
 
