@@ -1,3 +1,3 @@
 # Be sure to restart your server when you modify this file.
 
-AdoptAThing::Application.config.session_store :cookie_store, key: '_adopt-a-thing_session'
+Rails.application.config.session_store :cookie_store, key: '_adopt-a-thing_session'
