@@ -10,6 +10,6 @@ class CreateRailsAdminHistoriesTable < ActiveRecord::Migration
       t.timestamps
     end
 
-    add_index(:rails_admin_histories, [:item, :table, :month, :year], name: 'index_rails_admin_histories' )
+    add_index(:rails_admin_histories, [:item, :table, :month, :year], name: 'index_rails_admin_histories')
   end
 end
