@@ -4,7 +4,7 @@ class MainControllerTest < ActionController::TestCase
   include Devise::TestHelpers
 
   setup do
-    request.env["devise.mapping"] = Devise.mappings[:user]
+    request.env['devise.mapping'] = Devise.mappings[:user]
     @user = users(:erik)
   end
 
